bugfix: lazy load

+++ bot/src/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, BigInteger, String
 
 from src.database import BaseModel
 from src.lang import Language
@@ -6,7 +6,7 @@ from src.lang import Language
 
 class User(BaseModel):
     __tablename__ = 'user'
-    tg_id = Column(Integer, unique=True, nullable=False)
+    tg_id = Column(BigInteger, unique=True, nullable=False)
     tg_name = Column(String(255), nullable=False)
     tg_username = Column(String(255), nullable=True)
     language = Column(String(2), nullable=False, default="ru")
